OkHttp test refactoring
